More warnings

+++ src/wizards/MatWiz.cpp
@@ -122,7 +122,7 @@ wxString MatWiz::GetValue() {
       if (m_matrixType == MATRIX_SYMMETRIC && i > j)
         cmd += m_inputs[j * m_width + i]->GetValue();
       else if (m_matrixType == MATRIX_ANTISYMMETRIC && i > j)
-        cmd += wxS("-(") + m_inputs[j * m_width + i]->GetValue() + wxS(")");
+        cmd += wxS("-(") + m_inputs[static_cast<long>(j) * m_width + i]->GetValue() + wxS(")");
       else {
         wxString entry = m_inputs[static_cast<size_t>(i) * m_width + j]->GetValue();
         if (entry == wxEmptyString)
